+++ NEWS.md
@@ -1,3 +1,9 @@
+# goeveg 0.2.0
+
+- Added functionality to use frequencies in racurve()
+- Added functionality to label species in racurve()
+- Package checked and tested on OS X
+
 # goeveg 0.1.6
 
-First release
+- First CRAN release